Move function to utils rather than consts
